requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pgvector==0.3.6
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
loguru==0.7.3
Markdown==3.7
python-slugify==8.0.4
Unidecode==1.3.8
Jinja2==3.1.5
python-dotenv==1.0.1
aiofiles==24.1.0
itsdangerous==2.2.0
passlib[bcrypt]
asyncpg==0.30.0
redis==5.2.1
slowapi==0.1.9
alembic==1.14.0
PyYAML==6.0.2
SQLAlchemy==2.0.36
openai==1.58.1
starlette==0.41.3
psycopg2-binary==2.9.10
